Reorganisation of producers into base and user facing producers
