+++ Middleware/src/main/java/ca/concordia/encs/citydata/core/contracts/IDataStore.java
@@ -12,8 +12,6 @@ import java.util.UUID;
  * 
  * Refactoring made: changed the method's signature to use UUIDs instead of Strings and default methods to still handle
  * String IDs, set, get, delete
- * @author: Gabriel C. Ullmann, Minette Zongo
- * @date: 2025-02-22
  */
 public interface IDataStore<E> {
