+++ Middleware/src/main/java/ca/concordia/encs/citydata/producers/CKANMetadataProducer.java
@@ -9,6 +9,7 @@ import ca.concordia.encs.citydata.core.contracts.IOperation;
 import ca.concordia.encs.citydata.core.contracts.IProducer;
 import ca.concordia.encs.citydata.core.contracts.IRunner;
 import ca.concordia.encs.citydata.core.utils.RequestOptions;
+import ca.concordia.encs.citydata.producers.base.JSONProducer;
 
 /*This producer can connect to a CKAN instance and fetch either dataset or resource metadata.
  * Author: Gabriel C. Ullmann
