+++ Middleware/src/main/java/ca/concordia/encs/citydata/producers/GeometryProducer.java
@@ -3,6 +3,7 @@ package ca.concordia.encs.citydata.producers;
 import java.security.InvalidParameterException;
 
 import ca.concordia.encs.citydata.core.implementations.AbstractProducer;
+import ca.concordia.encs.citydata.producers.base.JSONProducer;
 import ca.concordia.encs.citydata.core.contracts.IOperation;
 import ca.concordia.encs.citydata.core.contracts.IProducer;
 import ca.concordia.encs.citydata.core.contracts.IRunner;
