+++ Middleware/src/test/java/ca/concordia/encs/citydata/EnergyConsumptionWithFilterTest.java
@@ -1,13 +1,15 @@
 package ca.concordia.encs.citydata;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.hamcrest.Matchers.containsString;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static org.hamcrest.Matchers.containsString;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.assertj.core.api.Assertions.assertThat;
 
+import java.lang.reflect.Field;
+import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.UUID;
 
@@ -23,74 +25,83 @@ import org.springframework.test.web.servlet.MvcResult;
 
 import ca.concordia.encs.citydata.core.configs.AppConfig;
 import ca.concordia.encs.citydata.datastores.InMemoryDataStore;
-import ca.concordia.encs.citydata.operations.StringFilterOperation;
 import ca.concordia.encs.citydata.producers.EnergyConsumptionProducer;
+import ca.concordia.encs.citydata.operations.StringFilterOperation;
+import ca.concordia.encs.citydata.PayloadFactory;
 import ca.concordia.encs.citydata.runners.SingleStepRunner;
 
 /***
- * This test validates the energy consumption data filtering through the API
- * endpoint and also directly through the producer component.
+ * This test validates the energy consumption data filtering through the API endpoint and also directly through 
+ * the producer component.
  * 
  * @author Minette Zongo M.
  * @date 2025-04-29
  */
 
+
 @SpringBootTest(classes = AppConfig.class)
 @AutoConfigureMockMvc
 @ComponentScan(basePackages = "ca.concordia.encs.citydata.core")
 public class EnergyConsumptionWithFilterTest {
-	@Autowired
-	private MockMvc mockMvc;
+    @Autowired
+    private MockMvc mockMvc;
 
-	private EnergyConsumptionProducer energyConsumptionProducer;
+    private EnergyConsumptionProducer energyConsumptionProducer;
 
-	@BeforeEach
-	void setUp() {
-		energyConsumptionProducer = new EnergyConsumptionProducer();
-	}
+    @BeforeEach
+    void setUp() {
+        energyConsumptionProducer = new EnergyConsumptionProducer();
+    }
+    
+    @Test
+    public void testEnergyConsumptionWithTimeFilter() throws Exception {
+    	
+        String jsonPayload = PayloadFactory.getExampleQuery("energyConsumptionWithFilter");
 
-	@Test
-	public void testEnergyConsumptionWithTimeFilter() throws Exception {
+        MvcResult mvcResult = mockMvc.perform(post("/apply/sync")
+                .contentType(MediaType.APPLICATION_JSON)
+                .content(jsonPayload))
+                .andExpect(status().isOk())
+                .andExpect(content().string(containsString("result")))
+                .andReturn();
+        
+        String responseContent = mvcResult.getResponse().getContentAsString();
+        System.out.println("API Response: " + responseContent);
+        
+        energyConsumptionProducer = new EnergyConsumptionProducer();
+        energyConsumptionProducer.setCity("montreal");
 
-		String jsonPayload = PayloadFactory.getExampleQuery("energyConsumptionWithFilter");
+        StringFilterOperation filterOperation = new StringFilterOperation();
+        filterOperation.setFilterBy("09:45:00");
+        energyConsumptionProducer.setOperation(filterOperation);
 
-		MvcResult mvcResult = mockMvc
-				.perform(post("/apply/sync").contentType(MediaType.APPLICATION_JSON).content(jsonPayload))
-				.andExpect(status().isOk()).andExpect(content().string(containsString("result"))).andReturn();
+        SingleStepRunner runner = new SingleStepRunner(energyConsumptionProducer);
+        UUID runnerId = UUID.randomUUID();
+        runner.setMetadata("id", runnerId.toString());
 
-		String responseContent = mvcResult.getResponse().getContentAsString();
+        Thread runnerThread = new Thread(() -> {
+            try {
+                System.out.println("Starting runner thread");
+                runner.runSteps();
+                System.out.println("Runner steps completed");
+            } catch (Exception e) {
+                System.err.println("Runner thread error: " + e.getMessage());
+                e.printStackTrace();
+            }
+        });
 
-		energyConsumptionProducer = new EnergyConsumptionProducer();
-		energyConsumptionProducer.setCity("montreal");
+        runnerThread.start();
+        runnerThread.join(); 
+        ArrayList<?> result = InMemoryDataStore.getInstance().get(runnerId).getResult();
 
-		StringFilterOperation filterOperation = new StringFilterOperation();
-		filterOperation.setFilterBy("09:45:00");
-		energyConsumptionProducer.setOperation(filterOperation);
+        assertNotNull(result, "Result should not be null");
+        assertThat(result).isNotEmpty();
 
-		SingleStepRunner runner = new SingleStepRunner(energyConsumptionProducer);
-		UUID runnerId = UUID.randomUUID();
-		runner.setMetadata("id", runnerId.toString());
-
-		Thread runnerThread = new Thread(() -> {
-			try {
-				runner.runSteps();
-			} catch (Exception e) {
-				e.printStackTrace();
-			}
-		});
-
-		runnerThread.start();
-		runnerThread.join();
-
-		ArrayList<?> result = InMemoryDataStore.getInstance().get(runnerId).getResult();
-
-		assertNotNull(result, "Result should not be null");
-		assertThat(result).isNotEmpty();
-
-		for (Object item : result) {
-			String itemString = item.toString();
-			assertTrue(itemString.contains("09:45:00"), "Filtered item should contain '09:45:00': " + itemString);
-		}
-	}
-
+        for (Object item : result) {
+            String itemString = item.toString();
+            assertTrue(itemString.contains("09:45:00"),
+                    "Filtered item should contain '09:45:00': " + itemString);
+        }
+    }
+       
 }
